Refactor tabs rendering.
