+++ app/lib/frontend/templates/views/shared/tabs.mustache
